fix: change nonce parameters from u8 to u64 to support larger nonce values
- Fix nonce parameter in TransactionArgs from u8 to u64
- Fix nested_safe_nonce parameter in TransactionArgs from u8 to u64
- Fix nonce parameter in tx_signing_hashes function from u8 to u64

This fixes the issue where nonce values > 255 would cause 'xxx is not in 0..=255' errors.
The u8 type was too restrictive for Safe transaction nonces which can grow much larger.

+++ crates/safe-hash/src/cli.rs
@@ -38,7 +38,7 @@ pub struct TransactionArgs {
 
     /// Transaction nonce of the safe address
     #[arg(short, long, required = true)]
-    pub nonce: u8,
+    pub nonce: u64,
 
     /// Address of the safe address
     #[arg(short = 's', long = "safe-address", required = true)]
@@ -70,7 +70,7 @@ pub struct TransactionArgs {
 
     /// Nested safe nonce
     #[arg(long)]
-    pub nested_safe_nonce: Option<u8>,
+    pub nested_safe_nonce: Option<u64>,
 
     #[arg(long, default_value_t = U256::ZERO)]
     pub safe_tx_gas: U256,
